+++ crates/release_plz/src/init/mod.rs
@@ -190,7 +190,9 @@ fn action_yaml(branch: &str, github_token: &str, owner: &str, trusted_publishing
     };
 
     let cargo_registry_token = if trusted_publishing {
-        "${{ steps.auth.outputs.token }}".to_string()
+        // release-plz will generate the token during the release process if needed,
+        // so we don't need to pass the token to the action.
+        String::new()
     } else {
         format!("${{{{ secrets.{CARGO_REGISTRY_TOKEN} }}}}")
     };
@@ -198,15 +200,6 @@ fn action_yaml(branch: &str, github_token: &str, owner: &str, trusted_publishing
     let id_token_permissions = if trusted_publishing {
         "
       id-token: write"
-    } else {
-        ""
-    };
-
-    let trusted_publishing_action = if trusted_publishing {
-        "
-      - name: Authenticate with crates.io
-        uses: rust-lang/crates-io-auth-action@v1
-        id: auth"
     } else {
         ""
     };
@@ -219,6 +212,16 @@ fn action_yaml(branch: &str, github_token: &str, owner: &str, trusted_publishing
     } else {
         // The crate might be private, so we add the token to the PR workflow.
         // If the crate is public, it won't hurt having it here. You can also remove it if you want.
+        format!(
+            "
+          CARGO_REGISTRY_TOKEN: {cargo_registry_token}"
+        )
+    };
+
+    let release_cargo_registry_token_env = if trusted_publishing {
+        // Omit the token in the release workflow as well when using trusted publishing.
+        "".to_string()
+    } else {
         format!(
             "
           CARGO_REGISTRY_TOKEN: {cargo_registry_token}"
@@ -241,19 +244,20 @@ jobs:
     permissions:
       contents: write{id_token_permissions}
     steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
+      - &checkout
+        name: Checkout repository
+        uses: actions/checkout@v5
         with:
           fetch-depth: 0{checkout_token_line}
-      - name: Install Rust toolchain
-        uses: dtolnay/rust-toolchain@stable{trusted_publishing_action}
+      - &install-rust
+        name: Install Rust toolchain
+        uses: dtolnay/rust-toolchain@stable
       - name: Run release-plz
         uses: release-plz/action@v0.5
         with:
           command: release
         env:
-          GITHUB_TOKEN: {github_token_secret}
-          CARGO_REGISTRY_TOKEN: {cargo_registry_token}
+          GITHUB_TOKEN: {github_token_secret}{release_cargo_registry_token_env}
 
   release-plz-pr:
     name: Release-plz PR
@@ -266,12 +270,8 @@ jobs:
       group: release-plz-${{{{ github.ref }}}}
       cancel-in-progress: false
     steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-        with:
-          fetch-depth: 0{checkout_token_line}
-      - name: Install Rust toolchain
-        uses: dtolnay/rust-toolchain@stable
+      - *checkout
+      - *install-rust
       - name: Run release-plz
         uses: release-plz/action@v0.5
         with:
@@ -336,11 +336,13 @@ mod tests {
                 permissions:
                   contents: write
                 steps:
-                  - name: Checkout repository
-                    uses: actions/checkout@v4
+                  - &checkout
+                    name: Checkout repository
+                    uses: actions/checkout@v5
                     with:
                       fetch-depth: 0
-                  - name: Install Rust toolchain
+                  - &install-rust
+                    name: Install Rust toolchain
                     uses: dtolnay/rust-toolchain@stable
                   - name: Run release-plz
                     uses: release-plz/action@v0.5
@@ -361,12 +363,8 @@ mod tests {
                   group: release-plz-${{ github.ref }}
                   cancel-in-progress: false
                 steps:
-                  - name: Checkout repository
-                    uses: actions/checkout@v4
-                    with:
-                      fetch-depth: 0
-                  - name: Install Rust toolchain
-                    uses: dtolnay/rust-toolchain@stable
+                  - *checkout
+                  - *install-rust
                   - name: Run release-plz
                     uses: release-plz/action@v0.5
                     with:
@@ -396,12 +394,14 @@ mod tests {
                 permissions:
                   contents: write
                 steps:
-                  - name: Checkout repository
-                    uses: actions/checkout@v4
+                  - &checkout
+                    name: Checkout repository
+                    uses: actions/checkout@v5
                     with:
                       fetch-depth: 0
                       token: ${{ secrets.RELEASE_PLZ_TOKEN }}
-                  - name: Install Rust toolchain
+                  - &install-rust
+                    name: Install Rust toolchain
                     uses: dtolnay/rust-toolchain@stable
                   - name: Run release-plz
                     uses: release-plz/action@v0.5
@@ -422,13 +422,8 @@ mod tests {
                   group: release-plz-${{ github.ref }}
                   cancel-in-progress: false
                 steps:
-                  - name: Checkout repository
-                    uses: actions/checkout@v4
-                    with:
-                      fetch-depth: 0
-                      token: ${{ secrets.RELEASE_PLZ_TOKEN }}
-                  - name: Install Rust toolchain
-                    uses: dtolnay/rust-toolchain@stable
+                  - *checkout
+                  - *install-rust
                   - name: Run release-plz
                     uses: release-plz/action@v0.5
                     with:
@@ -460,23 +455,21 @@ fn actions_yaml_string_with_trusted_publishing_is_correct() {
                   contents: write
                   id-token: write
                 steps:
-                  - name: Checkout repository
-                    uses: actions/checkout@v4
+                  - &checkout
+                    name: Checkout repository
+                    uses: actions/checkout@v5
                     with:
                       fetch-depth: 0
                       token: ${{ secrets.RELEASE_PLZ_TOKEN }}
-                  - name: Install Rust toolchain
+                  - &install-rust
+                    name: Install Rust toolchain
                     uses: dtolnay/rust-toolchain@stable
-                  - name: Authenticate with crates.io
-                    uses: rust-lang/crates-io-auth-action@v1
-                    id: auth
                   - name: Run release-plz
                     uses: release-plz/action@v0.5
                     with:
                       command: release
                     env:
                       GITHUB_TOKEN: ${{ secrets.RELEASE_PLZ_TOKEN }}
-                      CARGO_REGISTRY_TOKEN: ${{ steps.auth.outputs.token }}
 
               release-plz-pr:
                 name: Release-plz PR
@@ -489,13 +482,8 @@ fn actions_yaml_string_with_trusted_publishing_is_correct() {
                   group: release-plz-${{ github.ref }}
                   cancel-in-progress: false
                 steps:
-                  - name: Checkout repository
-                    uses: actions/checkout@v4
-                    with:
-                      fetch-depth: 0
-                      token: ${{ secrets.RELEASE_PLZ_TOKEN }}
-                  - name: Install Rust toolchain
-                    uses: dtolnay/rust-toolchain@stable
+                  - *checkout
+                  - *install-rust
                   - name: Run release-plz
                     uses: release-plz/action@v0.5
                     with: